Update app for generate JAR file

+++ src/test/java/com/kamil/servicExpert/service/ElementServiceTest.java
@@ -65,7 +65,7 @@ class ElementServiceTest {
 		// Then
         assertEquals(elementService.findById(id), Optional.of(element));
         assertEquals(elementService.findById(id).get().getElementName(), "nameOfElement");
-        assertEquals(elementService.findById(id).get().getElementPrice(), 20);
+        assertEquals(elementService.findById(id).get().getElementPrice(), BigDecimal.valueOf(20));
         verify(elementService, times(3)).findById(id);
 	}
